# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Button:
    properties:
      action:
        example: create
        type: string
      code:
        example: create
        type: string
      created_at:
        type: string
      id:
        example: 1
        type: integer
      menu_id:
        example: 1
        type: integer
      name:
        example: 创建用户
        type: string
      permission_code:
        example: user:create
        type: string
      updated_at:
        type: string
    type: object
  models.Menu:
    properties:
      button_association:
        example: false
        type: boolean
      component:
        example: '@/views/system/index'
        type: string
      created_at:
        type: string
      icon:
        example: setting
        type: string
      id:
        example: 1
        type: integer
      is_visible:
        example: true
        type: boolean
      meta:
        $ref: '#/definitions/models.MenuMeta'
      name:
        example: 系统管理
        type: string
      order:
        example: 1
        type: integer
      parent_id:
        example: 0
        type: integer
      path:
        example: /system
        type: string
      updated_at:
        type: string
    type: object
  models.MenuMeta:
    properties:
      icon:
        type: string
      title:
        type: string
    type: object
  models.Permission:
    properties:
      button_id:
        example: 1
        type: integer
      created_at:
        type: string
      id:
        example: 1
        type: integer
      menu_id:
        example: 1
        type: integer
      name:
        example: 查看仪表盘
        type: string
      permission_code:
        example: view_dashboard
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      type:
        example: menu
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        example: 系统管理员角色
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      id:
        example: 1
        type: integer
      password:
        example: password123
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        example: admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 租户中心系统的API文档
  title: 租户中心API
  version: "1.0"
paths:
...
  /api/buttons/menu/{menuId}:
    get:
      consumes:
      - application/json
      description: 获取指定菜单的所有按钮列表，需要管理员权限
      parameters:
      - description: 菜单ID
        in: path
        name: menuId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 按钮列表
          schema:
            items:
              $ref: '#/definitions/models.Button'
            type: array
        "400":
          description: 无效的菜单ID
          schema:
            type: object
        "500":
          description: 获取按钮列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 按钮管理
...
  /api/menus:
    get:
      consumes:
      - application/json
      description: 获取所有菜单的列表，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 菜单列表
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "500":
          description: 获取菜单列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新菜单，需要管理员权限
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 菜单创建成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 创建菜单失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
  /api/menus/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定菜单的信息，需要管理员权限
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 菜单信息更新成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 更新菜单信息失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
  /api/menus/{id}/permission:
    post:
      consumes:
      - application/json
      description: 为指定菜单绑定权限，需要管理员权限
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 权限绑定成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 绑定权限失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
  /api/menus/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定菜单的权限列表，需要管理员权限
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
        "400":
          description: 无效的菜单ID
          schema:
            type: object
        "500":
          description: 获取菜单权限列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
  /api/menus/parent/{parentId}:
    get:
      consumes:
      - application/json
      description: 获取指定父级菜单的子菜单列表，需要管理员权限
      parameters:
      - description: 父级菜单ID
        in: path
        name: parentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 子菜单列表
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
        "400":
          description: 无效的父级菜单ID
          schema:
            type: object
        "500":
          description: 获取子菜单列表失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 菜单管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"